Revert "Simplify auto_load_expand_dir_vars and remove substitute_path_component"
This reverts commit ea2fe2244e676c70032b7ad578eacb5b2f122482.

This commit was a refactoring to remove an xrealloc and simplify
utils.[ch].  However, it has a flaw -- it mishandles a substitution
like "$datadir/subdir".

I am backing out the patch in the interests of fixing the regression
before GDB 14.  It can be reinstated (with modifications) later if we
like.

Regression tested on x86-64 Fedora 36.